+++ src/core_modules/capture-core/components/WorkingLists/EventWorkingLists/types/apiTemplate.types.js
@@ -68,5 +68,6 @@ export type CommonQueryData = {|
     programId: string,
     orgUnitId: string,
     categories: ?Object,
-    programStageId?: ?string
+    programStageId?: ?string,
+    ouMode?: ?string,
 |};